(fix): Move formatPhoneNumber server-side
