+++ lib/api/helpers.js
@@ -196,36 +196,6 @@ export function getSlug(slugString) {
 }
 
 /**
- * @name formatPhoneNumber
- * @method
- * @memberof Helpers
- * @summary prepends country code to phone no. if required
- * @param  {String} phone the original phone no.
- * @param  {String} countryCode the country's code to which the phone no. belongs to
- * @return {Promise} that resolves to a string - the phone no. with country extension.
- */
-export async function formatPhoneNumber(phone, countryCode) {
-  try {
-    const { parseNumber, isValidNumber } = await import("libphonenumber-js");
-
-    // phone no. already has the country code attached
-    if (isValidNumber(phone)) {
-      return phone;
-    }
-
-    const parsed = parseNumber(phone, countryCode, { extended: true });
-    const { countryCallingCode, phone: phoneNumber } = parsed || {};
-    if (phoneNumber && countryCallingCode) {
-      // try attaching the country code to phone number
-      return `${countryCallingCode}${phoneNumber}`;
-    }
-    throw new ReactionError("invalid-parameter", `Incorrect format for phone number ${phone} with country ${countryCode}`);
-  } catch (error) {
-    throw new ReactionError("invalid-parameter", error);
-  }
-}
-
-/**
  * @name toCamelCase
  * @method
  * @memberof Helpers
